1.5.10a
- implement toggle details
- added it to logwizard form
- when turned on, the other msg_details need to be hidden
- IMPORTANT:
adding the details control on the Logwizard form slowed down
resizing/moving a LOT on debug mode
my take is that it's because of the many
splitter-container-inside-splitter-container, and perhaps it's combined
with the fact
that they are all using Dock=Fill.
also, description_ctrl has 6-embedded-into-one-another splitter
controls.
I'm hoping that this won't matter in release (about to test now)